Create q3.c
Sanvi is a very naughty kid. She can recognize numbers from 0 to 9. Her mother has to submit
a report to her team leader in next one hour. To engage the naughty Sanvi, she gives her 8 
balloons which has numbered printed on it. Sanvi is asked to put the balloons in a box, in such 
a way that the box shouldn’t contain duplicate balloons with respect to numbers printed on it. 
Help baby Sanvi to finish the task soon and start her mischiefs.