Create q6.c
Observe the following images, Given an image (a), Sanvi is not comfortable viewing it in her computer/tab. She wants a better viewing experience. Write a program that keeps Sanvi happy. (Hint-1. Image is represented as matrix; The input is a matrix and the output is a matrix as well.)
(Hint-2 Input matrix need to be transposed and rotated)