Create q7.c
Consider a string “R N S Institute of Technology”. People are fond of
abbreviations. So the output expected for above string is RNSIT. Write a code to implement
the same. Make your code generic so that it can abbreviate any given string say “British
Broadcasting Corporation” to BBC.
Note: Do not use Built-in libraries.