Create q8.c
Baby Sanvi’s mother is preparing her for monthly assessment. To assess Sanvi’s recognition
and counting ability, mother shows a English sentence say “a big black bear sat on a big black 
rug” and ask her to identify each letter and to tell the most frequent alphabet. In this case it’s a
with count 6. Automate this process by writing a program in C.
Note: Assume that the sentence will have only lower-case letters and no special characters.