Create q10.c
Doesn’t it represent stack of books? 
Each book has the credentials such as 
Book_Id (int)
Book_Name (char)
Book_Author (char)
Book_Price (float)
You are required to arrange the books in such a way that, only the book at the top can be 
removed, and new book can be added only on to the top of the existing book stack. Imagine 
these books are to be kept in a box. The maximum number of books the box can contain is 5. 
Write C routine to 
• insert book
• remove book
• display the existing book details
The program should also check the following constraints.
• if the box is full then, books can’t be added
• books can’t be removed if the box is empty
Note : Use Structures. Enhance the program with pointer to structure concept (Introduce -> operator)